+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Make some settings (e.g. frame rate) configurable through a TOML config file.  For
   more information see
   [documentation](https://open-dynamic-robot-initiative.github.io/trifinger_cameras/doc/configuration.html)
+- Script `tricamera_monitor_rate.py` for checking the actual frame rate (most meant for
+  debugging and testing during development)
 
 ### Removed
 - Obsolete script `verify_calibration.py`